fixde typo; move convert_quantize_dequantize to common
